split registry and extension into multiple files
