--- rust/otap-dataflow/crates/engine/src/capability/registry.rs
@@ -1,1779 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-//! Capability registry and resolution.
-//!
-//! [`CapabilityRegistry`] collects type-erased capability trait objects from
-//! extensions during the build phase. [`Capabilities`] is produced per-node
-//! by `resolve_bindings()` and provides the consumer API
-//! (`require_local`, `require_shared`, `optional_local`, `optional_shared`).
-//!
-//! TODO(extension-system): the pieces in this module are scaffolding —
-//! exercised only by the unit tests below. The engine build phase will
-//! populate the registry from extension factories, invoke
-//! `resolve_bindings()` per node, and use the `ConsumedTracker`'s
-//! `unconsumed_local` / `unconsumed_shared` methods to drop unused
-//! extension variants.
-
-use otap_df_config::ExtensionId;
-use std::any::{Any, TypeId};
-use std::cell::Cell;
-use std::collections::{HashMap, HashSet};
-use std::marker::PhantomData;
-use std::rc::Rc;
-
-/// Error type alias for capability operations.
-pub type Error = crate::error::Error;
-
-// ── Shared-capability factory trait ──────────────────────────────────────────
-
-/// Object-safe factory for the shared variant of a capability.
-///
-/// One concrete impl exists per `(capability, extension)` pair. The impl
-/// holds the extension instance by value (typically a `Clone + Send +
-/// 'static` type with `Arc`-wrapped shared state — see the
-/// [extension system architecture doc][arch]) and:
-///
-/// - [`clone_box`](Self::clone_box) duplicates the factory itself so
-///   each resolved per-node entry owns an independent `Box`. This is
-///   the `Box + Clone` idiom: `dyn Trait` is not `Clone`, so we thread
-///   cloning through an object-safe method.
-/// - [`produce_any`](Self::produce_any) mints a `Box<dyn shared::Trait>`
-///   for a consumer and type-erases it as `Box<dyn Any + Send>` for
-///   downcast at the [`Capabilities`] boundary.
-///
-/// "Factory" here follows the DI-container convention: a thing that
-/// hands you an instance on demand. The lifetime policy is an
-/// implementation detail of [`produce_any`] — today every capability
-/// clones a prototype (GoF Prototype pattern), tomorrow a capability
-/// could construct a fresh instance instead.
-///
-/// The trait is `Send`-only (not `Send + Sync`): registries are never
-/// shared across threads — they are cloned/owned per consumer — and
-/// `Box<T>: Send` does not require `T: Sync`. This matches the extension
-/// design doc's contract that shared extensions are `Clone + Send` with
-/// `Arc`-wrapped state.
-///
-/// [arch]: ../../../docs/extension-system-architecture.md
-#[doc(hidden)]
-pub trait SharedCapabilityFactory: Send {
-    /// Duplicate this factory. Each resolved per-node entry owns its
-    /// own `Box<dyn SharedCapabilityFactory>` produced via this method.
-    fn clone_box(&self) -> Box<dyn SharedCapabilityFactory>;
-
-    /// Produce a fresh shared trait object for a consumer. The returned
-    /// `Box<dyn Any + Send>` contains `Box<dyn shared::Trait>` — the
-    /// `require_shared` code downcasts to recover the concrete trait.
-    ///
-    /// # Why two `Box`es?
-    ///
-    /// The registry stores factories for many different capabilities in
-    /// one `HashMap`, so the factory trait must be object-safe and its
-    /// produce method must return a single uniform type — hence
-    /// [`Box<dyn Any + Send>`]. But `Any` requires its contents to be
-    /// `Sized`, and `dyn shared::Trait` is `!Sized`. Wrapping the trait
-    /// object in an inner `Box<dyn shared::Trait>` makes it `Sized` (a
-    /// fat pointer) so it can ride inside the `Any` envelope. The
-    /// outer `Box` is the uniform storage; the inner `Box` owns the
-    /// heap-allocated trait object.
-    fn produce_any(&self) -> Box<dyn Any + Send>;
-
-    /// Produce a fresh shared instance and wrap it as a local trait
-    /// object, returning `Rc<Rc<dyn local::Trait>>` type-erased to
-    /// `Rc<dyn Any>`. Used by [`resolve_bindings`] to populate the
-    /// per-node local slot when an extension provides only the shared
-    /// variant (`SharedAsLocal` fallback).
-    ///
-    /// This method exists on the factory — not on
-    /// [`ExtensionCapability`] — because the factory knows both the
-    /// capability type and the concrete `Shared` type at
-    /// monomorphization time. That lets the impl call
-    /// [`ExtensionCapability::wrap_shared_as_local`] directly, with no
-    /// runtime downcast and therefore no panic path.
-    ///
-    /// # Implementation contract
-    ///
-    /// Implementors must call `C::wrap_shared_as_local(produced)` and
-    /// wrap the result in an outer `Rc` so the value can be stored as
-    /// `Rc<dyn Any>`. The `#[capability]` proc macro will emit the
-    /// correct shape; today the only hand-written impls live in engine
-    /// crate tests.
-    ///
-    /// [`ExtensionCapability`]: super::ExtensionCapability
-    /// [`ExtensionCapability::wrap_shared_as_local`]: super::ExtensionCapability::wrap_shared_as_local
-    fn adapt_as_local_any(&self) -> Rc<dyn Any>;
-}
-
-/// Downcast a factory's [`produce_any`] output to `Box<C::Shared>`.
-///
-/// Hides the double-box erasure convention
-/// (`Box<Box<dyn Shared>>` stored as `Box<dyn Any + Send>`) that
-/// [`SharedCapabilityFactory::produce_any`] uses.
-///
-/// # Errors
-///
-/// Returns [`Error::InternalError`] if the factory produced a
-/// `Box<dyn Any + Send>` whose inner type is not `Box<C::Shared>`. This
-/// indicates a bug in a hand-rolled [`SharedCapabilityFactory`] impl
-/// that used the wrong inner box type. Macro-generated impls uphold
-/// the convention by construction.
-///
-/// [`produce_any`]: SharedCapabilityFactory::produce_any
-pub(crate) fn downcast_produced<C: super::ExtensionCapability>(
-    factory: &dyn SharedCapabilityFactory,
-) -> Result<Box<C::Shared>, Error> {
-    factory
-        .produce_any()
-        .downcast::<Box<C::Shared>>()
-        .map(|b| *b)
-        .map_err(|_| Error::InternalError {
-            message: format!(
-                "capability '{}': shared entry type mismatch (internal error)",
-                C::name(),
-            ),
-        })
-}
-
-/// A [`SharedCapabilityFactory`] implementing the **ClonePerConsumer**
-/// instance policy: a prototype extension value `E` is stored at
-/// registration time, and every consumer receives a fresh
-/// `Box<dyn C::Shared>` containing an independent clone of `E`.
-///
-/// Consumers do **not** share state — each holds its own copy.
-/// (If a capability needs shared mutable state across consumers, the
-/// `E` should internally hold an `Arc`/`Mutex`; `Clone` on `E` then
-/// clones the handle, not the state.)
-///
-/// Symmetric in name with [`ClonePerConsumerLocalFactory`]: the two
-/// encode the same registration-time policy (one prototype is stored;
-/// each consumer gets a clone of it) on the two ownership shapes
-/// (`Rc` local, `Box` shared).
-///
-/// The `to_shared` fn pointer is the per-`(C, E)` bridge that coerces
-/// an owned `E` into a `Box<dyn C::Shared>`; typically
-/// `|e| Box::new(e)` when `E: C::Shared`. It captures the coercion at
-/// construction so `produce_any` needs no runtime downcast.
-#[doc(hidden)]
-pub struct ClonePerConsumerSharedFactory<C: super::ExtensionCapability, E: Clone + Send + 'static> {
-    extension: E,
-    to_shared: fn(E) -> Box<C::Shared>,
-}
-
-impl<C: super::ExtensionCapability, E: Clone + Send + 'static> ClonePerConsumerSharedFactory<C, E> {
-    /// Construct a ClonePerConsumer shared factory. `to_shared` is the
-    /// per-`(C, E)` bridge that boxes an owned `E` as `Box<dyn C::Shared>`;
-    /// typically `|e| Box::new(e)` when `E: C::Shared`.
-    #[must_use]
-    pub fn new(extension: E, to_shared: fn(E) -> Box<C::Shared>) -> Self {
-        Self {
-            extension,
-            to_shared,
-        }
-    }
-}
-
-impl<C: super::ExtensionCapability, E: Clone + Send + 'static> SharedCapabilityFactory
-    for ClonePerConsumerSharedFactory<C, E>
-{
-    fn clone_box(&self) -> Box<dyn SharedCapabilityFactory> {
-        Box::new(ClonePerConsumerSharedFactory::<C, E> {
-            extension: self.extension.clone(),
-            to_shared: self.to_shared,
-        })
-    }
-    fn produce_any(&self) -> Box<dyn Any + Send> {
-        // Double-box convention: the inner `Box<C::Shared>` is a fat
-        // pointer (Sized) and can ride inside `Box<dyn Any + Send>`.
-        let shared: Box<C::Shared> = (self.to_shared)(self.extension.clone());
-        Box::new(shared)
-    }
-    fn adapt_as_local_any(&self) -> Rc<dyn Any> {
-        let shared: Box<C::Shared> = (self.to_shared)(self.extension.clone());
-        let rc_local = C::wrap_shared_as_local(shared);
-        Rc::new(rc_local)
-    }
-}
-
-/// A [`SharedCapabilityFactory`] implementing the **FreshPerConsumer**
-/// instance policy: a closure `F` is stored at registration time, and
-/// every consumer receives a freshly-constructed `Box<dyn C::Shared>`
-/// by invoking the closure.
-///
-/// Counterpart to [`ClonePerConsumerSharedFactory`]: where the Clone
-/// variant clones a prototype per consumer, this variant constructs a
-/// new instance per consumer. Useful when the extension has no
-/// meaningful `Clone` (or cloning would be semantically wrong — e.g.,
-/// buffers, file handles) but can be built on demand from captured
-/// configuration.
-///
-/// `F: Clone` is required so [`clone_box`](Self::clone_box) can hand
-/// per-node factories their own copy of the closure. Typical closures
-/// capture `Clone` configuration (`Arc<Config>`, `String`, etc.) and
-/// are naturally `Clone`.
-#[doc(hidden)]
-pub struct FreshPerConsumerSharedFactory<C: super::ExtensionCapability, F>
-where
-    F: Fn() -> Box<C::Shared> + Send + Clone + 'static,
-{
-    produce: F,
-    _phantom: PhantomData<fn() -> C>,
-}
-
-impl<C: super::ExtensionCapability, F> FreshPerConsumerSharedFactory<C, F>
-where
-    F: Fn() -> Box<C::Shared> + Send + Clone + 'static,
-{
-    /// Construct a FreshPerConsumer shared factory from a closure that
-    /// produces a fresh `Box<dyn C::Shared>` on each call.
-    #[must_use]
-    pub fn new(produce: F) -> Self {
-        Self {
-            produce,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-impl<C: super::ExtensionCapability, F> SharedCapabilityFactory
-    for FreshPerConsumerSharedFactory<C, F>
-where
-    F: Fn() -> Box<C::Shared> + Send + Clone + 'static,
-{
-    fn clone_box(&self) -> Box<dyn SharedCapabilityFactory> {
-        Box::new(FreshPerConsumerSharedFactory::<C, F> {
-            produce: self.produce.clone(),
-            _phantom: PhantomData,
-        })
-    }
-    fn produce_any(&self) -> Box<dyn Any + Send> {
-        let shared: Box<C::Shared> = (self.produce)();
-        Box::new(shared)
-    }
-    fn adapt_as_local_any(&self) -> Rc<dyn Any> {
-        let shared: Box<C::Shared> = (self.produce)();
-        let rc_local = C::wrap_shared_as_local(shared);
-        Rc::new(rc_local)
-    }
-}
-
-// ── Local-capability factory trait ───────────────────────────────────────────
-
-/// Object-safe factory for the local variant of a capability.
-///
-/// The local-side mirror of [`SharedCapabilityFactory`]. One concrete
-/// impl exists per `(capability, extension)` registration. The impl
-/// encodes the **instance policy** — how the registry supplies
-/// instances when consumers call [`Capabilities::require_local`]:
-///
-/// - [`ClonePerConsumerLocalFactory`] — all consumers receive an
-///   `Rc::clone` of a single cached instance (**ClonePerConsumer**
-///   policy; "clone" here is `Rc::clone`, so consumers observe a
-///   shared underlying object, mirroring the shared-side policy of
-///   handing out a clone per consumer). This is the only policy
-///   implemented today.
-/// - Future policies (e.g. fresh-per-consumer via a closure) will
-///   appear as additional impls of this trait; the registry, consumer
-///   API, and `resolve_bindings` will not change.
-///
-/// The trait is object-safe: entries store `Box<dyn
-/// LocalCapabilityFactory>` so multiple policies can coexist in one
-/// registry. Not `Send`: local factories hold `Rc`s and must never
-/// cross thread boundaries.
-///
-/// [`Capabilities::require_local`]: Capabilities::require_local
-#[doc(hidden)]
-pub trait LocalCapabilityFactory {
-    /// Duplicate this factory. Each resolved per-node entry owns its
-    /// own `Box<dyn LocalCapabilityFactory>` produced via this method.
-    /// Policies that cache an underlying `Rc` share it across clones
-    /// so every per-node clone points at the same instance.
-    fn clone_box(&self) -> Box<dyn LocalCapabilityFactory>;
-
-    /// Produce an `Rc<dyn local::Trait>` for a consumer, type-erased
-    /// as `Rc<dyn Any>`. [`Capabilities::require_local`] downcasts to
-    /// `Rc<C::Local>` to recover the concrete trait. Whether this
-    /// returns a cached clone or a freshly-built instance is a
-    /// policy-level choice (see the trait docs).
-    fn produce_any(&self) -> Rc<dyn Any>;
-}
-
-/// A [`LocalCapabilityFactory`] implementing the **ClonePerConsumer**
-/// instance policy: every consumer receives an `Rc::clone` of a single
-/// stored `Rc<dyn local::Trait>`.
-///
-/// Note on semantics: because cloning here is `Rc::clone` (not a deep
-/// data clone), consumers share the **same** underlying trait object —
-/// any interior mutability (`RefCell`, `Cell`, …) is visible across
-/// consumers. This matches the behavior of the pre-factory local
-/// registration and is symmetrical in *name* with the shared-side
-/// ClonePerConsumer policy, which hands out independently-cloned
-/// `Box<dyn shared::Trait>` instances. If fully-independent local
-/// instances are ever needed, a future `FreshPerConsumerLocalFactory`
-/// will cover that.
-///
-/// The `instance` field stores an `Rc<dyn Any>` whose inner type is
-/// `Rc<dyn local::Trait>`. The double-`Rc` is the local-side analogue
-/// of [`SharedCapabilityFactory`]'s double-`Box` convention: the outer
-/// `Rc` is the type-erased container; the inner `Rc<dyn Trait>` is
-/// what `require_local` downcasts to and clones out.
-#[doc(hidden)]
-pub struct ClonePerConsumerLocalFactory {
-    instance: Rc<dyn Any>,
-}
-
-impl ClonePerConsumerLocalFactory {
-    /// Wrap a type-erased `Rc<dyn Any>` (whose inner value must be
-    /// `Rc<dyn local::Trait>`) as a ClonePerConsumer local factory.
-    #[must_use]
-    pub fn new(instance: Rc<dyn Any>) -> Self {
-        Self { instance }
-    }
-}
-
-impl LocalCapabilityFactory for ClonePerConsumerLocalFactory {
-    fn clone_box(&self) -> Box<dyn LocalCapabilityFactory> {
-        Box::new(ClonePerConsumerLocalFactory {
-            instance: Rc::clone(&self.instance),
-        })
-    }
-    fn produce_any(&self) -> Rc<dyn Any> {
-        Rc::clone(&self.instance)
-    }
-}
-
-/// A [`LocalCapabilityFactory`] implementing the **FreshPerConsumer**
-/// instance policy: a closure `F` is stored at registration time, and
-/// every consumer receives a freshly-constructed `Rc<dyn Any>` (whose
-/// inner type must be `Rc<dyn local::Trait>`) by invoking the closure.
-///
-/// Counterpart to [`ClonePerConsumerLocalFactory`]: where the Clone
-/// variant shares one cached `Rc` across all consumers, this variant
-/// constructs a new instance per consumer. Consumers observe
-/// **independent** underlying trait objects — interior mutability is
-/// not shared.
-///
-/// `F: Clone` is required so [`clone_box`](Self::clone_box) can hand
-/// per-node factories their own copy of the closure.
-///
-/// The closure is responsible for producing the double-`Rc` shape
-/// `Rc<dyn Any>` containing `Rc<dyn local::Trait>` — the same
-/// convention as the stored form of
-/// [`ClonePerConsumerLocalFactory::instance`]. The extension-builder
-/// layer handles this wrapping when it constructs the factory.
-#[doc(hidden)]
-pub struct FreshPerConsumerLocalFactory<F>
-where
-    F: Fn() -> Rc<dyn Any> + Clone + 'static,
-{
-    produce: F,
-}
-
-impl<F> FreshPerConsumerLocalFactory<F>
-where
-    F: Fn() -> Rc<dyn Any> + Clone + 'static,
-{
-    /// Construct a FreshPerConsumer local factory from a closure that
-    /// produces `Rc<dyn Any>` wrapping `Rc<dyn local::Trait>` on each
-    /// call.
-    #[must_use]
-    pub fn new(produce: F) -> Self {
-        Self { produce }
-    }
-}
-
-impl<F> LocalCapabilityFactory for FreshPerConsumerLocalFactory<F>
-where
-    F: Fn() -> Rc<dyn Any> + Clone + 'static,
-{
-    fn clone_box(&self) -> Box<dyn LocalCapabilityFactory> {
-        Box::new(FreshPerConsumerLocalFactory {
-            produce: self.produce.clone(),
-        })
-    }
-    fn produce_any(&self) -> Rc<dyn Any> {
-        (self.produce)()
-    }
-}
-
-// ── Type-erased entries ──────────────────────────────────────────────────────
-
-/// A type-erased local (!Send) capability entry.
-///
-/// Holds a [`LocalCapabilityFactory`] whose concrete type encodes the
-/// instance policy for this `(capability, extension)` registration.
-/// Today the only impl is [`ClonePerConsumerLocalFactory`]; future
-/// policies will plug in as additional impls without changing this
-/// struct.
-///
-/// Extensions that only register a shared variant are handled by
-/// [`resolve_bindings`]: it invokes the shared factory's
-/// [`adapt_as_local_any`](SharedCapabilityFactory::adapt_as_local_any)
-/// and wraps the result in a [`ClonePerConsumerLocalFactory`] so the
-/// `SharedAsLocal` path flows through the same `LocalCapabilityFactory`
-/// pipeline as a native local registration.
-#[doc(hidden)]
-pub struct LocalCapabilityEntry {
-    /// The extension that provided this capability.
-    pub(crate) extension_id: ExtensionId,
-    /// Factory supplying local trait objects for this `(cap, ext)` pair.
-    pub(crate) factory: Box<dyn LocalCapabilityFactory>,
-}
-
-/// A type-erased shared (Send) capability entry.
-///
-/// [`factory`](Self::factory) is a per-`(capability, extension)` producer
-/// that can both clone itself and mint `Box<dyn shared::Trait>`
-/// instances. See [`SharedCapabilityFactory`].
-#[doc(hidden)]
-pub struct SharedCapabilityEntry {
-    /// The extension that provided this capability.
-    pub(crate) extension_id: ExtensionId,
-    /// Cloneable factory for shared capability trait objects.
-    pub(crate) factory: Box<dyn SharedCapabilityFactory>,
-}
-
-// ── CapabilityRegistry ───────────────────────────────────────────────────────
-
-/// Collects type-erased capability registrations from extensions during
-/// the pipeline build phase.
-///
-/// Keyed by `TypeId` of the zero-sized `ExtensionCapability` registration
-/// struct (e.g., `TypeId::of::<BearerTokenProvider>()`).
-///
-/// After all extensions have registered, call `resolve_bindings()` to
-/// produce per-node [`Capabilities`] structs.
-pub struct CapabilityRegistry {
-    /// Local entries keyed by (capability TypeId, extension name).
-    local: HashMap<TypeId, HashMap<ExtensionId, LocalCapabilityEntry>>,
-    /// Shared entries keyed by (capability TypeId, extension name).
-    shared: HashMap<TypeId, HashMap<ExtensionId, SharedCapabilityEntry>>,
-}
-
-impl CapabilityRegistry {
-    /// Creates an empty registry.
-    #[must_use]
-    pub fn new() -> Self {
-        CapabilityRegistry {
-            local: HashMap::new(),
-            shared: HashMap::new(),
-        }
-    }
-
-    /// Register a local capability entry for the given capability and extension.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`Error::InternalError`] if an entry already exists for
-    /// the same `(capability_id, entry.extension_id)` pair. Each
-    /// extension must register a given capability at most once; a
-    /// duplicate indicates an engine-side or extension-author bug.
-    pub fn register_local(
-        &mut self,
-        capability_id: TypeId,
-        entry: LocalCapabilityEntry,
-    ) -> Result<(), Error> {
-        let ext_id = entry.extension_id.clone();
-        let slot = self.local.entry(capability_id).or_default();
-        if slot.contains_key(&ext_id) {
-            return Err(Error::InternalError {
-                message: format!(
-                    "duplicate local capability registration: extension '{ext_id}' already registered capability TypeId {capability_id:?}",
-                ),
-            });
-        }
-        let _ = slot.insert(ext_id, entry);
-        Ok(())
-    }
-
-    /// Register a shared capability entry for the given capability and extension.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`Error::InternalError`] if an entry already exists for
-    /// the same `(capability_id, entry.extension_id)` pair. Each
-    /// extension must register a given capability at most once; a
-    /// duplicate indicates an engine-side or extension-author bug.
-    pub fn register_shared(
-        &mut self,
-        capability_id: TypeId,
-        entry: SharedCapabilityEntry,
-    ) -> Result<(), Error> {
-        let ext_id = entry.extension_id.clone();
-        let slot = self.shared.entry(capability_id).or_default();
-        if slot.contains_key(&ext_id) {
-            return Err(Error::InternalError {
-                message: format!(
-                    "duplicate shared capability registration: extension '{ext_id}' already registered capability TypeId {capability_id:?}",
-                ),
-            });
-        }
-        let _ = slot.insert(ext_id, entry);
-        Ok(())
-    }
-
-    /// Look up a local capability entry by capability type and extension name.
-    #[must_use]
-    pub fn get_local(
-        &self,
-        capability_id: &TypeId,
-        extension_id: &str,
-    ) -> Option<&LocalCapabilityEntry> {
-        self.local
-            .get(capability_id)
-            .and_then(|m| m.get(extension_id))
-    }
-
-    /// Look up a shared capability entry by capability type and extension name.
-    #[must_use]
-    pub fn get_shared(
-        &self,
-        capability_id: &TypeId,
-        extension_id: &str,
-    ) -> Option<&SharedCapabilityEntry> {
-        self.shared
-            .get(capability_id)
-            .and_then(|m| m.get(extension_id))
-    }
-
-    /// Returns `true` if any extension provides a **native** local entry
-    /// for this capability. A local binding is also reachable whenever
-    /// [`has_shared`](Self::has_shared) is true — shared registrations
-    /// are always adaptable via `SharedAsLocal`. For the composite
-    /// "can a node bind this as local?" predicate, use
-    /// `has_native_local(id) || has_shared(id)`.
-    #[must_use]
-    pub(crate) fn has_native_local(&self, capability_id: &TypeId) -> bool {
-        self.local
-            .get(capability_id)
-            .is_some_and(|m| !m.is_empty())
-    }
-
-    /// Returns `true` if any extension provides a shared entry for this capability.
-    #[must_use]
-    pub(crate) fn has_shared(&self, capability_id: &TypeId) -> bool {
-        self.shared
-            .get(capability_id)
-            .is_some_and(|m| !m.is_empty())
-    }
-}
-
-impl Default for CapabilityRegistry {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl std::fmt::Debug for CapabilityRegistry {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // The entry values are type-erased (`dyn Any` / `dyn SharedCapabilityFactory`)
-        // and can't be printed. Summarize the shape instead.
-        f.debug_struct("CapabilityRegistry")
-            .field("local_capabilities", &self.local.len())
-            .field("shared_capabilities", &self.shared.len())
-            .finish()
-    }
-}
-
-// ── Resolved entries (per-node) ──────────────────────────────────────────────
-
-/// A resolved local capability entry for a specific node.
-pub(crate) struct ResolvedLocalEntry {
-    /// Per-node local factory, produced by cloning the registry
-    /// entry's factory. The concrete impl encodes the instance policy
-    /// (today always [`ClonePerConsumerLocalFactory`]).
-    pub(crate) factory: Box<dyn LocalCapabilityFactory>,
-    /// Consumption flag for the underlying extension variant.
-    ///
-    /// - For native local bindings: the cell lives under the tracker's
-    ///   *local* bucket for the `(capability, extension)` pair.
-    /// - For `SharedAsLocal` bindings: the cell lives under the tracker's
-    ///   *shared* bucket — consuming the adapter is considered a use of
-    ///   the shared variant, because that's the only variant the
-    ///   extension actually provided. No phantom local entry is recorded.
-    pub(crate) consumed: Rc<Cell<bool>>,
-}
-
-/// A resolved shared capability entry for a specific node.
-pub(crate) struct ResolvedSharedEntry {
-    /// Per-node factory, produced by cloning the registry entry's factory.
-    /// Each node owns an independent `Box` (no cross-thread aliasing).
-    pub(crate) factory: Box<dyn SharedCapabilityFactory>,
-    /// Consumption flag, shared with the `ConsumedTracker`.
-    pub(crate) consumed: Rc<Cell<bool>>,
-}
-
-// ── Capabilities (consumer API) ──────────────────────────────────────────────
-
-/// Per-node capability bindings resolved from the [`CapabilityRegistry`].
-///
-/// Passed to node factories as `&Capabilities`. Provides type-safe
-/// access to extension capabilities via the [`ExtensionCapability`]
-/// sealed trait.
-///
-/// [`ExtensionCapability`]: super::ExtensionCapability
-pub struct Capabilities {
-    local: HashMap<TypeId, ResolvedLocalEntry>,
-    shared: HashMap<TypeId, ResolvedSharedEntry>,
-}
-
-impl Capabilities {
-    /// Creates a new `Capabilities` from resolved entries.
-    pub(crate) fn new(
-        local: HashMap<TypeId, ResolvedLocalEntry>,
-        shared: HashMap<TypeId, ResolvedSharedEntry>,
-    ) -> Self {
-        Capabilities { local, shared }
-    }
-
-    /// Creates an empty `Capabilities` (no bindings).
-    #[must_use]
-    pub fn empty() -> Self {
-        Capabilities {
-            local: HashMap::new(),
-            shared: HashMap::new(),
-        }
-    }
-
-    /// Resolve a **required** local capability.
-    ///
-    /// Returns `Rc<dyn C::Local>`. If the capability was registered by a
-    /// shared-only extension, the `SharedAsLocal` adapter is returned
-    /// transparently — the caller always gets a local trait object.
-    ///
-    /// # Errors
-    ///
-    /// - [`Error::ConfigError`] if no extension is bound to this
-    ///   capability for this node. The user must add a binding to the
-    ///   node's config.
-    /// - [`Error::InternalError`] on a type-erasure downcast mismatch
-    ///   (indicates a registry bug).
-    pub fn require_local<C: super::ExtensionCapability>(&self) -> Result<Rc<C::Local>, Error> {
-        let id = TypeId::of::<C>();
-        let entry = self.local.get(&id).ok_or_else(|| {
-            Error::ConfigError(Box::new(
-                otap_df_config::error::Error::InvalidUserConfig {
-                    error: format!(
-                        "required local capability '{}' not bound for this node",
-                        C::name(),
-                    ),
-                },
-            ))
-        })?;
-        let rc_any = entry.factory.produce_any();
-        let trait_object = rc_any
-            .downcast_ref::<Rc<C::Local>>()
-            .cloned()
-            .ok_or_else(|| Error::InternalError {
-                message: format!(
-                    "capability '{}': local entry type mismatch (internal error)",
-                    C::name(),
-                ),
-            })?;
-        entry.consumed.set(true);
-        Ok(trait_object)
-    }
-
-    /// Resolve a **required** shared capability.
-    ///
-    /// Returns `Box<dyn C::Shared>` — a fresh clone of the extension's
-    /// shared implementation.
-    ///
-    /// # Errors
-    ///
-    /// - [`Error::ConfigError`] if no extension is bound to this
-    ///   capability for this node. The user must add a binding to the
-    ///   node's config.
-    /// - [`Error::InternalError`] on a type-erasure downcast mismatch
-    ///   (indicates a registry bug).
-    pub fn require_shared<C: super::ExtensionCapability>(&self) -> Result<Box<C::Shared>, Error> {
-        let id = TypeId::of::<C>();
-        let entry = self.shared.get(&id).ok_or_else(|| {
-            Error::ConfigError(Box::new(
-                otap_df_config::error::Error::InvalidUserConfig {
-                    error: format!(
-                        "required shared capability '{}' not bound for this node",
-                        C::name(),
-                    ),
-                },
-            ))
-        })?;
-        let trait_object = downcast_produced::<C>(&*entry.factory)?;
-        entry.consumed.set(true);
-        Ok(trait_object)
-    }
-
-    /// Resolve an **optional** local capability.
-    ///
-    /// Returns `Some(Rc<dyn C::Local>)` if bound, `None` if the capability
-    /// was not configured for this node.
-    ///
-    /// # Panics
-    ///
-    /// Panics on type mismatch (internal bug — the stored entry doesn't
-    /// match the expected capability type).
-    #[must_use]
-    pub fn optional_local<C: super::ExtensionCapability>(&self) -> Option<Rc<C::Local>> {
-        let id = TypeId::of::<C>();
-        if !self.local.contains_key(&id) {
-            return None;
-        }
-        Some(
-            self.require_local::<C>().expect(
-                "BUG: capability entry exists but downcast failed — type mismatch in registry",
-            ),
-        )
-    }
-
-    /// Resolve an **optional** shared capability.
-    ///
-    /// Returns `Some(Box<dyn C::Shared>)` if bound, `None` if the capability
-    /// was not configured for this node.
-    ///
-    /// # Panics
-    ///
-    /// Panics on type mismatch (internal bug — the stored entry doesn't
-    /// match the expected capability type).
-    #[must_use]
-    pub fn optional_shared<C: super::ExtensionCapability>(&self) -> Option<Box<C::Shared>> {
-        let id = TypeId::of::<C>();
-        if !self.shared.contains_key(&id) {
-            return None;
-        }
-        Some(
-            self.require_shared::<C>().expect(
-                "BUG: capability entry exists but downcast failed — type mismatch in registry",
-            ),
-        )
-    }
-}
-
-impl std::fmt::Debug for Capabilities {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // Entry values are type-erased; only the shape is printable.
-        f.debug_struct("Capabilities")
-            .field("local_bindings", &self.local.len())
-            .field("shared_bindings", &self.shared.len())
-            .finish()
-    }
-}
-
-// ── ConsumedTracker ──────────────────────────────────────────────────────────
-
-/// Tracks which capability variants were consumed by node factories.
-///
-/// Created alongside per-node [`Capabilities`] during `resolve_bindings()`.
-/// After all node factories have run, the engine inspects this tracker to
-/// determine which extension variants are unused and can be dropped.
-///
-/// Keyed by `(capability TypeId, extension ID)` so multiple providers of
-/// the same capability are tracked independently. The `Rc<Cell<bool>>`
-/// for a given key is shared across all nodes that bind to that
-/// provider — once any of them consumes the capability the cell is set.
-pub(crate) struct ConsumedTracker {
-    local: HashMap<(TypeId, ExtensionId), ConsumedEntry>,
-    shared: HashMap<(TypeId, ExtensionId), ConsumedEntry>,
-}
-
-impl std::fmt::Debug for ConsumedTracker {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // `TypeId` is `Debug` but its output is opaque (`TypeId { t: ... }`);
-        // summarize counts instead. For detail, see `unconsumed_local` /
-        // `unconsumed_shared`, which return human-readable names.
-        f.debug_struct("ConsumedTracker")
-            .field("local_slots", &self.local.len())
-            .field("shared_slots", &self.shared.len())
-            .finish()
-    }
-}
-
-/// A single consumption tracking entry.
-pub(crate) struct ConsumedEntry {
-    /// Human-readable capability name (for warnings).
-    pub(crate) name: &'static str,
-    /// The extension that provides this capability.
-    pub(crate) extension_id: ExtensionId,
-    /// Shared flag — set to `true` when any node consumes this variant.
-    pub(crate) consumed: Rc<Cell<bool>>,
-}
-
-impl ConsumedTracker {
-    /// Creates an empty tracker.
-    pub(crate) fn new() -> Self {
-        ConsumedTracker {
-            local: HashMap::new(),
-            shared: HashMap::new(),
-        }
-    }
-
-    /// Returns the `Rc<Cell<bool>>` tracking local consumption for this
-    /// `(capability, extension)` pair, creating a fresh cell (initialized
-    /// to `false`) if none exists yet. This **registers** the consumer
-    /// slot; it does **not** mark the capability as consumed. The cell
-    /// is flipped to `true` later, only when a consumer calls
-    /// [`Capabilities::require_local`] / [`Capabilities::optional_local`].
-    pub(crate) fn ensure_local_consumer_slot(
-        &mut self,
-        capability_id: TypeId,
-        name: &'static str,
-        extension_id: ExtensionId,
-    ) -> Rc<Cell<bool>> {
-        let entry = self
-            .local
-            .entry((capability_id, extension_id.clone()))
-            .or_insert_with(|| ConsumedEntry {
-                name,
-                extension_id,
-                consumed: Rc::new(Cell::new(false)),
-            });
-        Rc::clone(&entry.consumed)
-    }
-
-    /// Returns the `Rc<Cell<bool>>` tracking shared consumption for this
-    /// `(capability, extension)` pair, creating a fresh cell (initialized
-    /// to `false`) if none exists yet. This **registers** the consumer
-    /// slot; it does **not** mark the capability as consumed. The cell
-    /// is flipped to `true` later, only when a consumer calls
-    /// [`Capabilities::require_shared`] / [`Capabilities::optional_shared`]
-    /// (or [`Capabilities::require_local`] / [`Capabilities::optional_local`]
-    /// for an extension that only provides a shared variant).
-    pub(crate) fn ensure_shared_consumer_slot(
-        &mut self,
-        capability_id: TypeId,
-        name: &'static str,
-        extension_id: ExtensionId,
-    ) -> Rc<Cell<bool>> {
-        let entry = self
-            .shared
-            .entry((capability_id, extension_id.clone()))
-            .or_insert_with(|| ConsumedEntry {
-                name,
-                extension_id,
-                consumed: Rc::new(Cell::new(false)),
-            });
-        Rc::clone(&entry.consumed)
-    }
-
-    /// Returns extension IDs whose local variant was never consumed.
-    /// Used by the engine to call `drop_local()` on those extensions.
-    pub(crate) fn unconsumed_local(&self) -> Vec<(ExtensionId, &'static str)> {
-        self.local
-            .values()
-            .filter(|e| !e.consumed.get())
-            .map(|e| (e.extension_id.clone(), e.name))
-            .collect()
-    }
-
-    /// Returns extension IDs whose shared variant was never consumed.
-    /// Used by the engine to call `drop_shared()` on those extensions.
-    pub(crate) fn unconsumed_shared(&self) -> Vec<(ExtensionId, &'static str)> {
-        self.shared
-            .values()
-            .filter(|e| !e.consumed.get())
-            .map(|e| (e.extension_id.clone(), e.name))
-            .collect()
-    }
-}
-
-// ── resolve_bindings ─────────────────────────────────────────────────────────
-
-/// Resolves a node's capability bindings against the registry.
-///
-/// For each `(capability_name, extension_name)` in the node's config:
-///
-/// 1. **Extension exists** — `extension_name` is in `known_extensions`.
-/// 2. **Known capability** — `capability_name` is in `KNOWN_CAPABILITIES`.
-/// 3. **Capability provided** — The registry has an entry for this capability.
-/// 4. **Specific extension provides it** — The registry entry was registered
-///    by `extension_name`.
-///
-/// On success, returns a [`Capabilities`] for the node and updates the
-/// `tracker` with consumption flags.
-///
-/// # Errors
-///
-/// Returns an error on a validation failure with a descriptive message.
-/// `bindings` is iterated in unspecified order (it's a `HashMap`), so
-/// when multiple bindings are invalid the specific one reported is not
-/// stable across runs.
-pub(crate) fn resolve_bindings(
-    bindings: &HashMap<otap_df_config::CapabilityId, ExtensionId>,
-    registry: &CapabilityRegistry,
-    known_extensions: &HashSet<ExtensionId>,
-    tracker: &mut ConsumedTracker,
-) -> Result<Capabilities, Error> {
-    // Build lookup from capability name → KnownCapability
-    let known_caps: HashMap<&str, &super::KnownCapability> = super::KNOWN_CAPABILITIES
-        .iter()
-        .map(|kc| (kc.name, kc))
-        .collect();
-
-    let mut local_entries: HashMap<TypeId, ResolvedLocalEntry> = HashMap::new();
-    let mut shared_entries: HashMap<TypeId, ResolvedSharedEntry> = HashMap::new();
-
-    for (cap_name, ext_name) in bindings {
-        let cap_name_str: &str = cap_name.as_ref();
-        let ext_name_str: &str = ext_name.as_ref();
-
-        // Step 1: Extension exists
-        if !known_extensions.contains(ext_name_str) {
-            return Err(Error::ConfigError(Box::new(
-                otap_df_config::error::Error::InvalidUserConfig {
-                    error: format!(
-                        "capability binding '{cap_name_str}': no extension named '{ext_name_str}' exists",
-                    ),
-                },
-            )));
-        }
-
-        // Step 2: Known capability type
-        let known_cap = known_caps.get(cap_name_str).ok_or_else(|| {
-            let known_names: Vec<&str> = known_caps.keys().copied().collect();
-            Error::ConfigError(Box::new(
-                otap_df_config::error::Error::InvalidUserConfig {
-                    error: format!(
-                        "unknown capability '{cap_name_str}'. Known capabilities: {known_names:?}",
-                    ),
-                },
-            ))
-        })?;
-
-        let cap_type_id = (known_cap.type_id)();
-
-        // Step 3: At least one extension must provide this capability
-        let has_native_local = registry.has_native_local(&cap_type_id);
-        let has_shared = registry.has_shared(&cap_type_id);
-        if !has_native_local && !has_shared {
-            return Err(Error::ConfigError(Box::new(
-                otap_df_config::error::Error::InvalidUserConfig {
-                    error: format!(
-                        "capability '{cap_name_str}': no loaded extension provides it",
-                    ),
-                },
-            )));
-        }
-
-        // Step 4: The specific named extension must provide it.
-        let local_entry = registry.get_local(&cap_type_id, ext_name_str);
-        let shared_entry = registry.get_shared(&cap_type_id, ext_name_str);
-
-        if local_entry.is_none() && shared_entry.is_none() {
-            return Err(Error::ConfigError(Box::new(
-                otap_df_config::error::Error::InvalidUserConfig {
-                    error: format!(
-                        "capability '{cap_name_str}': extension '{ext_name_str}' does not provide it",
-                    ),
-                },
-            )));
-        }
-
-        // Resolve local entry: prefer a native local registration; else,
-        // if the extension registered a shared entry, invoke the
-        // capability's `SharedAsLocal` adapter to build a fresh local
-        // wrapper around this node's own clone of the shared instance.
-        let native_local = local_entry;
-
-        if let Some(local_entry) = native_local {
-            let consumed = tracker.ensure_local_consumer_slot(
-                cap_type_id,
-                known_cap.name,
-                local_entry.extension_id.clone(),
-            );
-
-            let _ = local_entries.insert(
-                cap_type_id,
-                ResolvedLocalEntry {
-                    factory: local_entry.factory.clone_box(),
-                    consumed,
-                },
-            );
-        } else if let Some(shared_entry) = shared_entry {
-            // SharedAsLocal fallback: invoke the shared factory's
-            // `adapt_as_local_any` to mint a fresh local trait object,
-            // then wrap that `Rc<dyn Any>` in a
-            // `ClonePerConsumerLocalFactory`
-            // so the fallback flows through the same
-            // `LocalCapabilityFactory` path as a native local
-            // registration. Each node gets a fresh adapter
-            // (shared-across-components within the node once cached).
-            // Shared registrations are always adaptable to local — the
-            // shared factory's impl is the single source of truth and
-            // contains no downcast or panic path.
-            let rc_any = shared_entry.factory.adapt_as_local_any();
-            let local_factory: Box<dyn LocalCapabilityFactory> =
-                Box::new(ClonePerConsumerLocalFactory::new(rc_any));
-
-            // The extension only provided a shared variant — route the
-            // adapter consumer's cell under the shared bucket. The
-            // follow-up `// Resolve shared entry` block below also calls
-            // `ensure_shared_consumer_slot` for this `(cap, ext)` pair;
-            // that method is get-or-insert, so both users share the same
-            // `Rc<Cell<bool>>` and either one flipping marks the shared
-            // variant consumed. No phantom entry is created in
-            // `tracker.local`.
-            let consumed = tracker.ensure_shared_consumer_slot(
-                cap_type_id,
-                known_cap.name,
-                shared_entry.extension_id.clone(),
-            );
-
-            let _ = local_entries.insert(
-                cap_type_id,
-                ResolvedLocalEntry {
-                    factory: local_factory,
-                    consumed,
-                },
-            );
-        }
-
-        // Resolve shared entry. If the SharedAsLocal fallback above
-        // already called `ensure_shared_consumer_slot` for this
-        // `(cap, ext)` pair, this call is a no-op lookup that returns
-        // the same cell — get-or-insert idempotency keeps the two
-        // users pointing at one slot.
-        if let Some(shared_entry) = shared_entry {
-            let consumed = tracker.ensure_shared_consumer_slot(
-                cap_type_id,
-                known_cap.name,
-                shared_entry.extension_id.clone(),
-            );
-
-            let _ = shared_entries.insert(
-                cap_type_id,
-                ResolvedSharedEntry {
-                    factory: shared_entry.factory.clone_box(),
-                    consumed,
-                },
-            );
-        }
-    }
-
-    Ok(Capabilities::new(local_entries, shared_entries))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::any::TypeId;
-
-    // ── Hand-written test capability ─────────────────────────────────────
-    //
-    // This manually implements `ExtensionCapability` and registers a
-    // `KnownCapability` entry — the same interfaces that the `#[capability]`
-    // proc macro targets. We can't use the macro here because proc macros
-    // cannot be invoked from within the crate that hosts their generated
-    // target paths (`crate::capability::*`).
-    //
-    // These tests verify the **registry infrastructure** (register, resolve,
-    // consume, SharedAsLocal adapter). Update the hand-written impls below if `ExtensionCapability`,
-    // `KnownCapability`, or the registry API changes.
-
-    /// A minimal test capability trait (local version).
-    trait TestCapLocal {
-        fn value(&self) -> &str;
-    }
-
-    /// A minimal test capability trait (shared version).
-    trait TestCapShared: Send {
-        fn value(&self) -> &str;
-    }
-
-    /// Zero-sized registration struct for `TestCap`.
-    struct TestCap;
-
-    impl super::super::private::Sealed for TestCap {}
-
-    impl super::super::ExtensionCapability for TestCap {
-        const NAME: &'static str = "test_cap";
-        type Local = dyn TestCapLocal;
-        type Shared = dyn TestCapShared;
-
-        fn wrap_shared_as_local(shared: Box<Self::Shared>) -> Rc<Self::Local> {
-            struct Adapter(Box<dyn TestCapShared>);
-            impl TestCapLocal for Adapter {
-                fn value(&self) -> &str {
-                    self.0.value()
-                }
-            }
-            Rc::new(Adapter(shared))
-        }
-    }
-
-    /// Register TestCap in KNOWN_CAPABILITIES for resolve_bindings tests.
-    #[allow(unsafe_code)]
-    #[linkme::distributed_slice(super::super::KNOWN_CAPABILITIES)]
-    #[linkme(crate = linkme)]
-    static _TEST_CAP: super::super::KnownCapability = super::super::KnownCapability {
-        name: "test_cap",
-        description: "Test capability for unit tests",
-        type_id: || TypeId::of::<TestCap>(),
-    };
-
-    // ── Test implementations ─────────────────────────────────────────────
-
-    #[derive(Clone)]
-    struct SharedImpl(&'static str);
-    impl TestCapShared for SharedImpl {
-        fn value(&self) -> &str {
-            self.0
-        }
-    }
-
-    struct LocalImpl(&'static str);
-    impl TestCapLocal for LocalImpl {
-        fn value(&self) -> &str {
-            self.0
-        }
-    }
-
-    // ── Helpers ──────────────────────────────────────────────────────────
-
-    // Bridge fn: turns an owned `SharedImpl` into a `Box<dyn TestCapShared>`.
-    // This is the per-`(cap, ext)` coercion that macro-generated
-    // registration code will emit for real capabilities; tests use it
-    // directly to construct `ClonePerConsumerSharedFactory<TestCap, SharedImpl>`.
-    fn shared_impl_as_box(s: SharedImpl) -> Box<dyn TestCapShared> {
-        Box::new(s)
-    }
-
-    fn test_factory(val: &'static str) -> ClonePerConsumerSharedFactory<TestCap, SharedImpl> {
-        ClonePerConsumerSharedFactory::<TestCap, SharedImpl>::new(
-            SharedImpl(val),
-            shared_impl_as_box,
-        )
-    }
-
-    fn register_shared(registry: &mut CapabilityRegistry, ext_id: &'static str, val: &'static str) {
-        let ext_id: ExtensionId = ext_id.into();
-        registry
-            .register_shared(
-                TypeId::of::<TestCap>(),
-                SharedCapabilityEntry {
-                    extension_id: ext_id,
-                    factory: Box::new(test_factory(val)),
-                },
-            )
-            .unwrap();
-    }
-
-    fn register_local(registry: &mut CapabilityRegistry, ext_id: &'static str, val: &'static str) {
-        let ext_id: ExtensionId = ext_id.into();
-        let rc_local: Rc<dyn TestCapLocal> = Rc::new(LocalImpl(val));
-        let instance: Rc<dyn Any> = Rc::new(rc_local);
-        registry
-            .register_local(
-                TypeId::of::<TestCap>(),
-                LocalCapabilityEntry {
-                    extension_id: ext_id,
-                    factory: Box::new(ClonePerConsumerLocalFactory::new(instance)),
-                },
-            )
-            .unwrap();
-    }
-
-    fn bindings(
-        cap: &'static str,
-        ext: &'static str,
-    ) -> HashMap<otap_df_config::CapabilityId, ExtensionId> {
-        let mut m = HashMap::new();
-        let _ = m.insert(cap.into(), ext.into());
-        m
-    }
-
-    fn known_exts(names: &[&'static str]) -> HashSet<ExtensionId> {
-        names.iter().map(|n| (*n).into()).collect()
-    }
-
-    // ── Tests ────────────────────────────────────────────────────────────
-
-    #[test]
-    fn test_registry_register_and_get() {
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-a", "hello");
-        assert!(reg.get_shared(&TypeId::of::<TestCap>(), "ext-a").is_some());
-        assert!(reg.get_shared(&TypeId::of::<TestCap>(), "ext-b").is_none());
-        assert!(reg.get_local(&TypeId::of::<TestCap>(), "ext-a").is_none());
-    }
-
-    #[test]
-    fn test_resolve_bindings_shared_only() {
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-a", "shared-val");
-
-        let mut tracker = ConsumedTracker::new();
-        let caps = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-
-        // require_shared works
-        let shared = caps.require_shared::<TestCap>().unwrap();
-        assert_eq!(shared.value(), "shared-val");
-
-        // require_local works via SharedAsLocal adapter
-        let local = caps.require_local::<TestCap>().unwrap();
-        assert_eq!(local.value(), "shared-val");
-    }
-
-    #[test]
-    fn test_resolve_bindings_local_only() {
-        let mut reg = CapabilityRegistry::new();
-        register_local(&mut reg, "ext-a", "local-val");
-
-        let mut tracker = ConsumedTracker::new();
-        let caps = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-
-        let local = caps.require_local::<TestCap>().unwrap();
-        assert_eq!(local.value(), "local-val");
-
-        // shared is not available
-        assert!(caps.require_shared::<TestCap>().is_err());
-    }
-
-    #[test]
-    fn test_resolve_bindings_step1_unknown_extension() {
-        let reg = CapabilityRegistry::new();
-        let mut tracker = ConsumedTracker::new();
-        let result = resolve_bindings(
-            &bindings("test_cap", "nonexistent"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        );
-        assert!(result.is_err());
-        let msg = format!("{}", result.err().unwrap());
-        assert!(msg.contains("nonexistent"), "error: {msg}");
-    }
-
-    #[test]
-    fn test_resolve_bindings_step2_unknown_capability() {
-        let reg = CapabilityRegistry::new();
-        let mut tracker = ConsumedTracker::new();
-        let result = resolve_bindings(
-            &bindings("totally_unknown_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        );
-        assert!(result.is_err());
-        let msg = format!("{}", result.err().unwrap());
-        assert!(msg.contains("unknown capability"), "error: {msg}");
-    }
-
-    #[test]
-    fn test_resolve_bindings_step3_not_provided() {
-        // Extension exists, capability is known, but no extension provides it.
-        let reg = CapabilityRegistry::new();
-        let mut tracker = ConsumedTracker::new();
-        let result = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        );
-        assert!(result.is_err());
-        let msg = format!("{}", result.err().unwrap());
-        assert!(msg.contains("no loaded extension provides"), "error: {msg}");
-    }
-
-    #[test]
-    fn test_resolve_bindings_step4_wrong_extension() {
-        // ext-b provides test_cap, but binding says ext-a.
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-b", "val");
-
-        let mut tracker = ConsumedTracker::new();
-        let result = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a", "ext-b"]),
-            &mut tracker,
-        );
-        assert!(result.is_err());
-        let msg = format!("{}", result.err().unwrap());
-        assert!(msg.contains("does not provide"), "error: {msg}");
-    }
-
-    #[test]
-    fn test_consumed_tracking_shared() {
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-a", "val");
-
-        let mut tracker = ConsumedTracker::new();
-        let caps = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-
-        // Not consumed yet
-        assert_eq!(tracker.unconsumed_shared().len(), 1);
-
-        // Consume
-        let _ = caps.require_shared::<TestCap>().unwrap();
-        assert!(tracker.unconsumed_shared().is_empty());
-    }
-
-    #[test]
-    fn test_consumed_tracking_local_marks_shared_via_adapter() {
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-a", "val");
-
-        let mut tracker = ConsumedTracker::new();
-        let caps = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-
-        // Consume via local adapter
-        let _ = caps.require_local::<TestCap>().unwrap();
-
-        // The extension only registered a shared variant, so there is no
-        // native local tracker entry for this (cap, ext) pair — the
-        // SharedAsLocal adapter's consumption counts as shared use.
-        assert!(tracker.unconsumed_local().is_empty());
-        assert!(
-            tracker.unconsumed_shared().is_empty(),
-            "consuming SharedAsLocal adapter must mark shared variant consumed",
-        );
-    }
-
-    #[test]
-    fn test_unconsumed_tracking() {
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-a", "val");
-
-        let mut tracker = ConsumedTracker::new();
-        let _caps = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-
-        // Not consumed — should appear in unconsumed
-        let unconsumed = tracker.unconsumed_shared();
-        assert_eq!(unconsumed.len(), 1);
-        assert_eq!(unconsumed[0].0.as_ref(), "ext-a");
-        assert_eq!(unconsumed[0].1, "test_cap");
-    }
-
-    #[test]
-    fn test_optional_returns_none_when_not_bound() {
-        let caps = Capabilities::empty();
-        assert!(caps.optional_local::<TestCap>().is_none());
-        assert!(caps.optional_shared::<TestCap>().is_none());
-    }
-
-    #[test]
-    fn test_extension_capabilities_none() {
-        let ec = super::super::ExtensionCapabilities::none();
-        assert!(ec.shared.is_empty());
-        assert!(ec.local.is_empty());
-    }
-
-    #[test]
-    fn test_extension_capabilities_shared_only() {
-        let ec = super::super::ExtensionCapabilities {
-            shared: &["bearer_token_provider"],
-            local: &[],
-        };
-        assert_eq!(ec.shared, &["bearer_token_provider"]);
-        assert!(ec.local.is_empty());
-    }
-
-    #[test]
-    fn test_known_capabilities_contains_test_cap() {
-        let found = super::super::KNOWN_CAPABILITIES
-            .iter()
-            .any(|kc| kc.name == "test_cap");
-        assert!(found, "test_cap should be in KNOWN_CAPABILITIES");
-    }
-
-    #[test]
-    fn test_multiple_providers_same_capability() {
-        // Two extensions both provide test_cap with different values.
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-a", "value-a");
-        register_shared(&mut reg, "ext-b", "value-b");
-
-        // Both are accessible by (cap, ext) key.
-        assert!(reg.get_shared(&TypeId::of::<TestCap>(), "ext-a").is_some());
-        assert!(reg.get_shared(&TypeId::of::<TestCap>(), "ext-b").is_some());
-
-        // Node bound to ext-a gets ext-a's value.
-        let mut tracker = ConsumedTracker::new();
-        let caps_a = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a", "ext-b"]),
-            &mut tracker,
-        )
-        .unwrap();
-        let shared_a = caps_a.require_shared::<TestCap>().unwrap();
-        assert_eq!(shared_a.value(), "value-a");
-
-        // Node bound to ext-b gets ext-b's value.
-        let mut tracker = ConsumedTracker::new();
-        let caps_b = resolve_bindings(
-            &bindings("test_cap", "ext-b"),
-            &reg,
-            &known_exts(&["ext-a", "ext-b"]),
-            &mut tracker,
-        )
-        .unwrap();
-        let shared_b = caps_b.require_shared::<TestCap>().unwrap();
-        assert_eq!(shared_b.value(), "value-b");
-    }
-
-    /// Regression: the `ConsumedTracker` must not lose track of a node's
-    /// consumption when a second node resolves the same capability.
-    ///
-    /// Scenario: two nodes bind `test_cap` to `ext-a` (local). Node A
-    /// consumes it, node B does not. The tracker should still report the
-    /// capability as consumed (because at least one node consumed it) —
-    /// otherwise the engine will drop an extension variant that's actually
-    /// in use.
-    #[test]
-    fn test_consumed_tracking_persists_across_nodes_local() {
-        let mut reg = CapabilityRegistry::new();
-        register_local(&mut reg, "ext-a", "val");
-
-        let mut tracker = ConsumedTracker::new();
-
-        // Node A resolves + consumes (local).
-        let caps_a = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-        let _ = caps_a.require_local::<TestCap>().unwrap();
-
-        // Node B resolves but does NOT consume.
-        let _caps_b = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-
-        // At least one node consumed it — should not appear in unconsumed.
-        assert!(
-            tracker.unconsumed_local().is_empty(),
-            "unconsumed_local should be empty but got {:?}",
-            tracker.unconsumed_local()
-        );
-    }
-
-    /// Same regression for shared variant (already works via get-or-insert,
-    /// but worth locking in behavior).
-    #[test]
-    fn test_consumed_tracking_persists_across_nodes_shared() {
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-a", "val");
-
-        let mut tracker = ConsumedTracker::new();
-
-        let caps_a = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-        let _ = caps_a.require_shared::<TestCap>().unwrap();
-
-        let _caps_b = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-
-        assert!(tracker.unconsumed_shared().is_empty());
-    }
-
-    /// Regression for Option C: each node resolving a `SharedAsLocal`
-    /// binding receives its own clone of the shared extension instance.
-    /// This preserves per-caller-fresh semantics that a shared impl may
-    /// rely on (e.g. per-caller mutable state) — the adapter is not
-    /// pre-built once and shared across nodes.
-    #[test]
-    fn test_shared_as_local_builds_fresh_adapter_per_node() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-        static CLONE_COUNT: AtomicUsize = AtomicUsize::new(0);
-
-        // Register a shared impl whose factory bumps a counter every
-        // time the extension is cloned.
-        CLONE_COUNT.store(0, Ordering::SeqCst);
-        let mut reg = CapabilityRegistry::new();
-        let ext_id: ExtensionId = "ext-a".into();
-
-        #[derive(Clone)]
-        struct CountingFactory;
-        impl SharedCapabilityFactory for CountingFactory {
-            fn clone_box(&self) -> Box<dyn SharedCapabilityFactory> {
-                Box::new(self.clone())
-            }
-            fn produce_any(&self) -> Box<dyn Any + Send> {
-                let _ = CLONE_COUNT.fetch_add(1, Ordering::SeqCst);
-                let shared: Box<dyn TestCapShared> = Box::new(SharedImpl("val"));
-                Box::new(shared)
-            }
-            fn adapt_as_local_any(&self) -> Rc<dyn Any> {
-                let _ = CLONE_COUNT.fetch_add(1, Ordering::SeqCst);
-                let shared: Box<dyn TestCapShared> = Box::new(SharedImpl("val"));
-                let rc_local =
-                    <TestCap as super::super::ExtensionCapability>::wrap_shared_as_local(shared);
-                Rc::new(rc_local)
-            }
-        }
-
-        reg.register_shared(
-            TypeId::of::<TestCap>(),
-            SharedCapabilityEntry {
-                extension_id: ext_id,
-                factory: Box::new(CountingFactory),
-            },
-        )
-        .unwrap();
-
-        // The adapter must not run at registration time — work is deferred
-        // to resolve_bindings so each node gets a fresh shared clone.
-        assert_eq!(CLONE_COUNT.load(Ordering::SeqCst), 0);
-
-        // Two nodes bind the capability; each should get its own clone.
-        let mut tracker = ConsumedTracker::new();
-        let _caps_a = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-        let _caps_b = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-
-        assert_eq!(CLONE_COUNT.load(Ordering::SeqCst), 2);
-    }
-
-    /// Duplicate registrations indicate a programmer bug — the registry
-    /// must reject them loudly rather than silently overwriting.
-    #[test]
-    fn test_register_local_rejects_duplicate() {
-        let mut reg = CapabilityRegistry::new();
-        register_local(&mut reg, "ext-a", "v1");
-
-        let rc_local: Rc<dyn TestCapLocal> = Rc::new(LocalImpl("v2"));
-        let instance: Rc<dyn Any> = Rc::new(rc_local);
-        let err = reg
-            .register_local(
-                TypeId::of::<TestCap>(),
-                LocalCapabilityEntry {
-                    extension_id: "ext-a".into(),
-                    factory: Box::new(ClonePerConsumerLocalFactory::new(instance)),
-                },
-            )
-            .unwrap_err();
-        let msg = format!("{err}");
-        assert!(msg.contains("duplicate"), "error: {msg}");
-        assert!(msg.contains("ext-a"), "error: {msg}");
-    }
-
-    #[test]
-    fn test_register_shared_rejects_duplicate() {
-        let mut reg = CapabilityRegistry::new();
-        register_shared(&mut reg, "ext-a", "v1");
-
-        let err = reg
-            .register_shared(
-                TypeId::of::<TestCap>(),
-                SharedCapabilityEntry {
-                    extension_id: "ext-a".into(),
-                    factory: Box::new(test_factory("v2")),
-                },
-            )
-            .unwrap_err();
-        let msg = format!("{err}");
-        assert!(msg.contains("duplicate"), "error: {msg}");
-        assert!(msg.contains("ext-a"), "error: {msg}");
-    }
-
-    // ── FreshPerConsumer factory tests ───────────────────────────────────
-
-    /// Each call to `produce_any` on a fresh-factory mints a new
-    /// instance: the counter captured by the closure increments,
-    /// proving independent construction.
-    #[test]
-    fn test_fresh_shared_factory_produces_independent_instances() {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let counter = Arc::new(AtomicUsize::new(0));
-        let counter_c = Arc::clone(&counter);
-        let fac = FreshPerConsumerSharedFactory::<TestCap, _>::new(move || {
-            let n = counter_c.fetch_add(1, Ordering::SeqCst);
-            // Leak a &'static str via Box is overkill; just use a fixed value —
-            // the assertion is on call count, not per-call payload distinctness.
-            let _ = n;
-            Box::new(SharedImpl("fresh")) as Box<dyn TestCapShared>
-        });
-
-        let _ = fac.produce_any();
-        let _ = fac.produce_any();
-        let _ = fac.produce_any();
-        assert_eq!(counter.load(Ordering::SeqCst), 3);
-    }
-
-    /// `clone_box` must duplicate the factory so per-node resolved
-    /// entries can each mint their own instances.
-    #[test]
-    fn test_fresh_shared_factory_clone_box_preserves_closure() {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let counter = Arc::new(AtomicUsize::new(0));
-        let counter_c = Arc::clone(&counter);
-        let fac = FreshPerConsumerSharedFactory::<TestCap, _>::new(move || {
-            let _ = counter_c.fetch_add(1, Ordering::SeqCst);
-            Box::new(SharedImpl("fresh")) as Box<dyn TestCapShared>
-        });
-
-        let cloned = fac.clone_box();
-        let _ = fac.produce_any();
-        let _ = cloned.produce_any();
-        assert_eq!(counter.load(Ordering::SeqCst), 2);
-    }
-
-    /// `adapt_as_local_any` for a fresh shared factory must invoke the
-    /// closure and route through `wrap_shared_as_local`, yielding a
-    /// downcastable `Rc<dyn TestCapLocal>`.
-    #[test]
-    fn test_fresh_shared_factory_adapt_as_local() {
-        let fac = FreshPerConsumerSharedFactory::<TestCap, _>::new(|| {
-            Box::new(SharedImpl("fresh-via-local")) as Box<dyn TestCapShared>
-        });
-        let any_rc = fac.adapt_as_local_any();
-        let rc_local = any_rc.downcast::<Rc<dyn TestCapLocal>>().unwrap();
-        assert_eq!(rc_local.value(), "fresh-via-local");
-    }
-
-    /// Each call to `produce_any` on a fresh local factory yields a
-    /// *different* underlying `Rc<dyn TestCapLocal>` — verified by
-    /// comparing raw pointers of the downcast trait objects.
-    #[test]
-    fn test_fresh_local_factory_produces_independent_instances() {
-        let fac = FreshPerConsumerLocalFactory::new(|| {
-            let rc_local: Rc<dyn TestCapLocal> = Rc::new(LocalImpl("fresh"));
-            Rc::new(rc_local) as Rc<dyn Any>
-        });
-
-        let a = fac.produce_any();
-        let b = fac.produce_any();
-        let a_inner = a.downcast::<Rc<dyn TestCapLocal>>().unwrap();
-        let b_inner = b.downcast::<Rc<dyn TestCapLocal>>().unwrap();
-        assert!(!Rc::ptr_eq(&a_inner, &b_inner));
-        assert_eq!(a_inner.value(), "fresh");
-        assert_eq!(b_inner.value(), "fresh");
-    }
-
-    /// `clone_box` on a fresh local factory duplicates the closure;
-    /// both the original and the clone still produce independent
-    /// instances on each call.
-    #[test]
-    fn test_fresh_local_factory_clone_box() {
-        let fac = FreshPerConsumerLocalFactory::new(|| {
-            let rc_local: Rc<dyn TestCapLocal> = Rc::new(LocalImpl("fresh-clone"));
-            Rc::new(rc_local) as Rc<dyn Any>
-        });
-
-        let cloned = fac.clone_box();
-        let a = fac.produce_any();
-        let b = cloned.produce_any();
-        let a_inner = a.downcast::<Rc<dyn TestCapLocal>>().unwrap();
-        let b_inner = b.downcast::<Rc<dyn TestCapLocal>>().unwrap();
-        assert!(!Rc::ptr_eq(&a_inner, &b_inner));
-    }
-
-    /// A fresh shared factory plugs into `SharedCapabilityEntry` and
-    /// flows through `resolve_bindings` / `require_shared` the same
-    /// way a clone factory does.
-    #[test]
-    fn test_fresh_shared_factory_via_registry() {
-        let mut reg = CapabilityRegistry::new();
-        reg.register_shared(
-            TypeId::of::<TestCap>(),
-            SharedCapabilityEntry {
-                extension_id: "ext-a".into(),
-                factory: Box::new(FreshPerConsumerSharedFactory::<TestCap, _>::new(|| {
-                    Box::new(SharedImpl("fresh-reg")) as Box<dyn TestCapShared>
-                })),
-            },
-        )
-        .unwrap();
-
-        let mut tracker = ConsumedTracker::new();
-        let caps = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-        let s = caps.require_shared::<TestCap>().unwrap();
-        assert_eq!(s.value(), "fresh-reg");
-    }
-
-    /// A fresh local factory plugs into `LocalCapabilityEntry` and
-    /// flows through `resolve_bindings` / `require_local` the same
-    /// way a clone factory does.
-    #[test]
-    fn test_fresh_local_factory_via_registry() {
-        let mut reg = CapabilityRegistry::new();
-        reg.register_local(
-            TypeId::of::<TestCap>(),
-            LocalCapabilityEntry {
-                extension_id: "ext-a".into(),
-                factory: Box::new(FreshPerConsumerLocalFactory::new(|| {
-                    let rc_local: Rc<dyn TestCapLocal> = Rc::new(LocalImpl("fresh-local-reg"));
-                    Rc::new(rc_local) as Rc<dyn Any>
-                })),
-            },
-        )
-        .unwrap();
-
-        let mut tracker = ConsumedTracker::new();
-        let caps = resolve_bindings(
-            &bindings("test_cap", "ext-a"),
-            &reg,
-            &known_exts(&["ext-a"]),
-            &mut tracker,
-        )
-        .unwrap();
-        let l = caps.require_local::<TestCap>().unwrap();
-        assert_eq!(l.value(), "fresh-local-reg");
-    }
-}
